# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "u-cafe",
        "description": "Documentation for u-cafe admin dashboard and main site",
        "contact": {
            "email": "s.murodjonov@uzinfocom.uz"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/order-column": {
            "get": {
                "tags": [
                    "OrderColumns"
                ],
                "summary": "Get list of OrderColumn",
                "description": "Returns list of OrderColumn",
                "operationId": "indexOrderColumn",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderColumns"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "OrderColumns"
                ],
                "summary": "Insert a new OrderColumn",
                "description": "Insert a new OrderColumn",
                "operationId": "storeOrderColumn",
                "requestBody": {
                    "description": "OrderColumn to create",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "$ref": "#/components/schemas/OrderColumn"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OrderColumn created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/OrderColumn"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation exception"
                    }
                }
            }
        },
        "/api/order-column/{ordercolumn_id}": {
            "get": {
                "tags": [
                    "OrderColumns"
                ],
                "summary": "Show a OrderColumn from his Id",
                "description": "Show a OrderColumn from his Id",
                "operationId": "showOrderColumn",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/OrderColumn--id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/OrderColumn"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "OrderColumn not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "OrderColumns"
                ],
                "summary": "Delete a OrderColumn",
                "description": "Delete a OrderColumn",
                "operationId": "destroyOrderColumn",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/OrderColumn--id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/OrderColumn"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "OrderColumn not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "OrderColumns"
                ],
                "summary": "Update an existing OrderColumn",
                "description": "Update an existing OrderColumn",
                "operationId": "updateOrderColumn",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/OrderColumn--id"
                    }
                ],
                "requestBody": {
                    "description": "OrderColumn to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "$ref": "#/components/schemas/OrderColumn"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/OrderColumn"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "OrderColumn not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "OrderColumn": {
                "title": "OrderColumn",
                "description": "OrderColumn model",
                "required": [],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id of OrderColumn",
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "title": "name",
                        "description": "name of OrderColumn",
                        "type": "string",
                        "example": "Macbook Pro"
                    },
                    "sku": {
                        "title": "sku",
                        "description": "sku of OrderColumn",
                        "type": "string",
                        "example": "MCBPRO2022"
                    },
                    "price": {
                        "title": "price",
                        "description": "price of OrderColumn",
                        "type": "integer",
                        "example": "99"
                    },
                    "created_at": {
                        "title": "created_at",
                        "type": "dateTime",
                        "readOnly": "true",
                        "example": "2022-07-04T02:41:42.336Z"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "type": "dateTime",
                        "readOnly": "true",
                        "example": "2022-07-04T02:41:42.336Z"
                    }
                },
                "type": "object"
            },
            "OrderColumns": {
                "title": "OrderColumns list",
                "properties": {
                    "data": {
                        "title": "data",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderColumn"
                        }
                    },
                    "first_page_url": {
                        "title": "first_page_url",
                        "type": "string",
                        "example": "http://localhost:8080/api/merchant-customers?page=1"
                    },
                    "last_page_url": {
                        "title": "last_page_url",
                        "type": "string",
                        "example": "http://localhost:8080/api/merchant-customers?page=3"
                    },
                    "prev_page_url": {
                        "title": "prev_page_url",
                        "type": "string",
                        "example": "null"
                    },
                    "next_page_url": {
                        "title": "next_page_url",
                        "type": "string",
                        "example": "http://localhost:8080/api/merchant-customers?page=2"
                    },
                    "current_page": {
                        "title": "current_page",
                        "type": "integer",
                        "example": "1"
                    },
                    "from": {
                        "title": "from",
                        "type": "integer",
                        "example": "1"
                    },
                    "last_page": {
                        "title": "last_page",
                        "type": "integer",
                        "example": "3"
                    },
                    "path": {
                        "title": "path",
                        "type": "string",
                        "example": "http://localhost:8080/api/merchant-customers"
                    },
                    "per_page": {
                        "title": "per_page",
                        "type": "integer",
                        "example": "1"
                    },
                    "total": {
                        "title": "total",
                        "type": "integer",
                        "example": "3"
                    },
                    "links": {
                        "title": "links",
                        "type": "array",
                        "items": {
                            "properties": {
                                "url": {
                                    "title": "url",
                                    "type": "string",
                                    "example": "http://localhost:8080/api/merchant-customers?page=2"
                                },
                                "label": {
                                    "title": "label",
                                    "type": "string",
                                    "example": "1"
                                },
                                "active": {
                                    "title": "active",
                                    "type": "boolean",
                                    "example": "true"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
...
            "OrderColumn--id": {
                "name": "OrderColumn_id",
                "in": "path",
                "description": "Id of OrderColumn",
                "required": true,
                "schema": {
                    "type": "integer",
                    "example": "1"
                }
            },
...
        }
    }
}